repos: probe local clone for refspec first
when using a branch or tag name as a refspec it helps to resolve the
refspec on the local clone first.
This allows local commits, which haven't been pushed to remote yet -
otherwise those would be overwritten with each run of `kas build`.
if ref-parse output is not matching the refspec it is likely a branch or
tag name that is need for checkout.
Using --force-checkout overrides that decision and always uses remote.

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>